refactor runtime config inferences
